perf(Block): remove the repeated message sending to server
To verify information of a block, the verify functions need to send
messages to server for requesting more information about this block or
pre block. Some requests are repeated which might increase the burden of
server. By saving some information already requsted, the client won't
need to send too many repeated requests to server and reduce the size of
log.